requirements.txt
wilds>=2.0
torch>=2.0
torchvision>=0.15
Pillow>=9.0
numpy>=1.24
scipy>=1.10
scikit-learn>=1.3
matplotlib>=3.7
tqdm>=4.65
pandas>=2.0
joblib>=1.2
